Heroku, Postman (server examples)

+++ Bamazon/schema.sql
@@ -8,6 +8,6 @@ CREATE TABLE products (
   product_name VARCHAR(100) NOT NULL,
   department_name VARCHAR(100) NOT NULL,
   price DECIMAL(10,2) NOT NULL,
-  stock_quantity DECIMAL(20,0) NOT NULL,
+  stock_quantity INTEGER(20) NOT NULL,
   PRIMARY KEY (id)
 );
